restructured the app - NEEDS CHECKING to make sure all works
