+++ Services/shak_decacher/main.js
@@ -1,5 +1,5 @@
-import {JIRA_TICKET_TEMPLATE_TASK, JIRA_TICKET_TEMPLATE_BUG} from './JIRA_TEMPLATES.js'
-import {callContentScript} from './background.js'
+import {JIRA_TICKET_TEMPLATE_TASK, JIRA_TICKET_TEMPLATE_BUG} from '../jira_injector/templates.js'
+
 import {SHAK_DECACHE_MATCHES, SHAK_FAIL_MESSAGE, SHAK_SUCCESS_MESSAGE, SHAK_URL} from './constants.js'
 
 let shakDecacheButton,
@@ -10,37 +10,26 @@ let shakDecacheButton,
 	jiraInjectionBugButton,
 	jiraInjectionTaskButton = null
 
-document.addEventListener('DOMContentLoaded', function () {
-	shakDecacheButton = document.querySelector('#shak_decache_button')
-	shakDecacheButtonSpinner = document.querySelector('#shak_decache_button_spinner')
-	shakDecacheMessage = document.querySelector('#shak_decache_message')
-	shakDecacheMessageResult = document.querySelector('#shak_decache_message_result')
-	shakDecacheMessageContent = document.querySelector('#shak_decache_message_content')
-	jiraInjectionBugButton = document.querySelector('#jira_injection_bug')
-	jiraInjectionTaskButton = document.querySelector('#jira_injection_task')
+export default () => {
+	document.addEventListener('DOMContentLoaded', function () {
+		shakDecacheButton = document.querySelector('#shak_decache_button')
+		shakDecacheButtonSpinner = document.querySelector('#shak_decache_button_spinner')
+		shakDecacheMessage = document.querySelector('#shak_decache_message')
+		shakDecacheMessageResult = document.querySelector('#shak_decache_message_result')
+		shakDecacheMessageContent = document.querySelector('#shak_decache_message_content')
+		jiraInjectionBugButton = document.querySelector('#jira_injection_bug')
+		jiraInjectionTaskButton = document.querySelector('#jira_injection_task')
 
-	shakDecacheButton.addEventListener('click', () => {
-		clearShakDecacheMessage()
-		hideShakDecacheMessage()
-		showSpinner()
-		decacheCurrentURL()
+		shakDecacheButton.addEventListener('click', () => {
+			clearShakDecacheMessage()
+			hideShakDecacheMessage()
+			showSpinner()
+			decacheCurrentURL()
+		})
+
+
 	})
-
-	jiraInjectionBugButton.addEventListener('click', async () =>
-		callContentScript({
-			action: 'JIRA_TEMPLATE',
-			jiraTemplateType: 'BUG',
-			jiraTemplateContent: JIRA_TICKET_TEMPLATE_BUG
-		})
-	)
-	jiraInjectionTaskButton.addEventListener('click', async () =>
-		callContentScript({
-			action: 'JIRA_TEMPLATE',
-			jiraTemplateType: 'TASK',
-			jiraTemplateContent: JIRA_TICKET_TEMPLATE_TASK
-		})
-	)
-})
+}
 
 const decacheCurrentURL = async () => {
 	chrome.tabs.query({active: true, currentWindow: true}, async tabs => {
